+++ src/main/java/mapreduce/matrix/Coordinate.java
@@ -1,4 +1,4 @@
 package mapreduce.matrix;
 
-public record Coordinate(int i, int j, double value) {
+public record Coordinate(int i, int j, int value) {
 }
