+++ src/main/java/mapreduce/operators/matrixmultiplication/DenseMatrixMultiplication.java
@@ -10,10 +10,10 @@ public class DenseMatrixMultiplication implements MatrixMultiplication {
     public Matrix multiply(Matrix matrix_a, Matrix matrix_b) {
         DenseMatrix a = (DenseMatrix) matrix_a;
         DenseMatrix b = (DenseMatrix) matrix_b;
-        double[][] c = new double[a.size()][b.size()];
+        int[][] c = new int[a.size()][b.size()];
         for (int i = 0; i < a.size(); i++) {
             for (int k = 0; k < b.size(); k++) {
-                double aik = a.get(i, k);
+                int aik = a.get(i, k);
                 for (int j = 0; j < a.size();j++) {
                     c[i][j] += aik * b.get(k, j);
                 }
